[type/__postfix*] Consolidate OS compatibility checks in __postfix
All __postfix_* types requires the __postfix/ object. The postconf utilities
make no difference between the operating systems.  The only requirement is that
the binaries are in $PATH which it might not be for some operating systems.

So, if any operating system checks should be done, they should be implemented in
the __postfix type and checks should be done for all types before adding a new
OS to type/__postfix/manifest.